Fix parsing to not rely on hardcoded entity expansions
JMdict uses internal entity references as enums. This causes me great grief.
Unfortunately C# does not help as it does not seem to have any way to
control the expansion of internal entities (as opposed to external ones)
outside of either prohibiting it (breaking parsing), or having it to be
default.

So unless we fix it, we have to parse the serializer's output after the
expansion. We used to map the expanded values back to enums, but this
approach is broken, as the during the last 2 years, the expanded values
have changed a lot more times often than the entity names themselves
(which happened once in that time, with cop-da -> cop)

So what this approach does is essentially grabbing the DTD along while
actually parsing the XML file, parsing the DTD entity type declarations
ourselves, and creating an inverse mapping from expansions back to
entity names, in order to undo the work entity expansion is doing.